Feita uma verificação para saber se o arquivo é terminado em '.db' pra n da exeption

+++ LauncherGUI/Launcher.cs
@@ -28,18 +28,24 @@ namespace LauncherGUI
 
         private void BtLocal_Click(object sender, EventArgs e)
         {
-            openFileDialog1.ShowDialog();
+            fdDBFinder.ShowDialog();
         }
 
-        private void OpenFileDialog1_FileOk(object sender, CancelEventArgs e)
+        private void fdDBFinder_FileOk(object sender, CancelEventArgs e)
         {
-            txtLocal.Text = openFileDialog1.FileName;
-            if (File.Exists(nomeArquivo))
+            if(!fdDBFinder.FileName.EndsWith(".db"))
             {
-                File.Delete(nomeArquivo);
+                MessageBox.Show("Essa não é um db válida");
+            } else
+            {
+               txtLocal.Text = fdDBFinder.FileName;
+               if (File.Exists(nomeArquivo))
+               {
+                   File.Delete(nomeArquivo);
+               }
+               File.Create(nomeArquivo).Close();
+               File.WriteAllText(nomeArquivo, txtLocal.Text);
             }
-            File.Create(nomeArquivo).Close();
-            File.WriteAllText(nomeArquivo, txtLocal.Text);
         }
 
         private void BtIniciar_Click(object sender, EventArgs e)
@@ -53,7 +59,6 @@ namespace LauncherGUI
             } catch (Exception erro)
             {
                 MessageBox.Show($"Não foi possivel iniciar o bot: {erro}");
-                return;
             }
         }
     }
